back-end/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
torch==2.0.0
torchvision==0.15.1
Pillow==10.1.0
numpy==1.24.3
gunicorn==20.1.0
flask-cors==3.0.10